+++ CHANGELOG.md
@@ -30,6 +30,8 @@ Release NEXT:
    * JsonPath4K 2.3.0 (with proper Kotlin 2.0.20 support)
    * Signum 3.7.0 (only dependency updates to align everything, no alignments in code)
  * Add `KeyStoreMaterial` to JVM target for convenience
+ - Update implementation of [OpenID for Verifiable Credential Issuance](https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html) to draft 14 from 2024-08-21
+ - Move some fields from `IssuerMetadata` to `OAuth2AuthorizationServerMetadata` to match the semantics
 
 Release 4.1.2:
  * In `OidcSiopVerifier` add parameter `nonceService` to externalize creation and validation of nonces, e.g. for deployments in load-balanced environments